+++ lib/yahl7/v2.rb
@@ -1,6 +1,11 @@
 # frozen_string_literal: true
 
-require_relative 'v2/segment_field_names'
+require_relative 'v2/alias_field_names'
+require_relative 'v2/alias_person_name'
+
+Dir[File.join(__dir__, 'v2', 'error', '*.rb')].sort.each { |f| require f }
+Dir[File.join(__dir__, 'v2', 'data_type', '*.rb')].sort.each { |f| require f }
+
 require_relative 'v2/date_time'
 require_relative 'v2/formatter'
 require_relative 'v2/parser'