+++ lib/yahl7/v2/error/invalid_format_error.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module YAHL7
+  module V2
+    module Error
+      class InvalidFormatError < StandardError
+      end
+    end
+  end
+end