+++ lib/yahl7/v2/segment/msh.rb
@@ -4,7 +4,7 @@ module YAHL7
   module V2
     class Segment
       class MSH < YAHL7::V2::Segment
-        include YAHL7::V2::SegmentFieldNames
+        include YAHL7::V2::AliasFieldNames
 
         define_field_names({
                              field_separator: 1,